pyproject.toml
[project]
name = "agentic-langraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.4",
    "langchain-tavily>=0.2.4",
    "langgraph>=0.4.10",
    "langsmith>=0.4.2",
    "python-dotenv>=1.1.1",
]
